Backing change made in revision 18372 for "z" and "digest" compression methods. These will likely give the same problem that arose with the "bz" method and required it to be reverted.
Implemented alternative method of supressing warnings due to hides overloaded virtual
function. This just declares that the class is "using" the method with the same name
(but different argument types) from the base class. We don't actually use it, but it
lets the compiler know that we know that it is there and our declaration using different
argument types is not a mistake